# packages/shared-types/openapi.json
{"openapi": "3.1.0", "info": {"title": "Time Internet Retention API", "version": "0.9.0"}, "paths": {"/api/health": {"get": {"tags": ["health"], "summary": "Health", "description": "Liveness probe; also reports the active LLM mode (mock|gemini).", "operationId": "health_api_health_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HealthResponse"}}}}}}}, "/api/customers": {"get": {"tags": ["customers"], "summary": "List Customers", "operationId": "list_customers_api_customers_get", "parameters": [{"name": "search", "in": "query", "required": false, "schema": {"anyOf": [{"type": "string"}, {"type": "null"}], "description": "matches name or customer id", "title": "Search"}, "description": "matches name or customer id"}, {"name": "plan", "in": "query", "required": false, "schema": {"anyOf": [{"type": "string"}, {"type": "null"}], "description": "filter by plan id", "title": "Plan"}, "description": "filter by plan id"}, {"name": "sort", "in": "query", "required": false, "schema": {"enum": ["tenure", "avg_gb", "contract_end_date"], "type": "string", "default": "contract_end_date", "title": "Sort"}}, {"name": "order", "in": "query", "required": false, "schema": {"enum": ["asc", "desc"], "type": "string", "default": "asc", "title": "Order"}}, {"name": "page", "in": "query", "required": false, "schema": {"type": "integer", "minimum": 1, "default": 1, "title": "Page"}}, {"name": "page_size", "in": "query", "required": false, "schema": {"type": "integer", "maximum": 100, "minimum": 1, "default": 20, "title": "Page Size"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Page_CustomerSummary_"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/customers/{customer_id}": {"get": {"tags": ["customers"], "summary": "Get Customer", "operationId": "get_customer_api_customers__customer_id__get", "parameters": [{"name": "customer_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Customer Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CustomerDetail"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/dashboard/summary": {"get": {"tags": ["dashboard"], "summary": "Summary", "operationId": "summary_api_dashboard_summary_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/DashboardSummary"}}}}}}}, "/api/customers/{customer_id}/pitch": {"post": {"tags": ["pitches"], "summary": "Generate Pitch", "operationId": "generate_pitch_api_customers__customer_id__pitch_post", "parameters": [{"name": "customer_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Customer Id"}}], "requestBody": {"content": {"application/json": {"schema": {"anyOf": [{"$ref": "#/components/schemas/PitchGenerateRequest"}, {"type": "null"}], "title": "Body"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}}, "components": {"schemas": {"CustomerDetail": {"properties": {"id": {"type": "string", "title": "Id"}, "name": {"type": "string", "title": "Name"}, "email": {"type": "string", "title": "Email"}, "phone": {"type": "string", "title": "Phone"}, "region": {"type": "string", "title": "Region"}, "current_plan": {"$ref": "#/components/schemas/PlanRef"}, "monthly_price": {"type": "integer", "title": "Monthly Price"}, "tenure_months": {"type": "integer", "title": "Tenure Months"}, "contract_end_date": {"type": "string", "format": "date", "title": "Contract End Date"}, "usage_archetype": {"type": "string", "title": "Usage Archetype"}, "avg_monthly_gb": {"type": "integer", "title": "Avg Monthly Gb"}, "last_month_gb": {"type": "integer", "title": "Last Month Gb"}, "usage_history": {"items": {"$ref": "#/components/schemas/UsagePoint"}, "type": "array", "title": "Usage History"}, "offer_ladder": {"$ref": "#/components/schemas/OfferLadder"}, "latest_pitch": {"anyOf": [{"$ref": "#/components/schemas/PitchRead"}, {"type": "null"}]}}, "type": "object", "required": ["id", "name", "email", "phone", "region", "current_plan", "monthly_price", "tenure_months", "contract_end_date", "usage_archetype", "avg_monthly_gb", "last_month_gb", "usage_history", "offer_ladder"], "title": "CustomerDetail", "description": "Full detail: customer info + usage history + offer ladder + latest pitch."}, "CustomerSummary": {"properties": {"id": {"type": "string", "title": "Id"}, "name": {"type": "string", "title": "Name"}, "current_plan": {"$ref": "#/components/schemas/PlanRef"}, "tenure_months": {"type": "integer", "title": "Tenure Months"}, "avg_monthly_gb": {"type": "integer", "title": "Avg Monthly Gb"}, "contract_end_date": {"type": "string", "format": "date", "title": "Contract End Date"}, "latest_pitch_status": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Latest Pitch Status"}}, "type": "object", "required": ["id", "name", "current_plan", "tenure_months", "avg_monthly_gb", "contract_end_date"], "title": "CustomerSummary", "description": "Row shape for the customer list."}, "DashboardSummary": {"properties": {"expiring_this_month": {"type": "integer", "title": "Expiring This Month"}, "by_tier": {"items": {"$ref": "#/components/schemas/TierCount"}, "type": "array", "title": "By Tier"}}, "type": "object", "required": ["expiring_this_month", "by_tier"], "title": "DashboardSummary"}, "HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "HealthResponse": {"properties": {"status": {"type": "string", "title": "Status"}, "llm_mode": {"type": "string", "title": "Llm Mode"}}, "type": "object", "required": ["status", "llm_mode"], "title": "HealthResponse"}, "OfferLadder": {"properties": {"rungs": {"items": {"$ref": "#/components/schemas/OfferRung"}, "type": "array", "title": "Rungs"}, "recommended": {"type": "string", "enum": ["retain", "value_upgrade", "upsell"], "title": "Recommended"}}, "type": "object", "required": ["rungs", "recommended"], "title": "OfferLadder"}, "OfferRung": {"properties": {"type": {"type": "string", "enum": ["retain", "value_upgrade", "upsell"], "title": "Type"}, "target_plan": {"$ref": "#/components/schemas/PlanRef"}, "monthly_price": {"type": "integer", "title": "Monthly Price"}, "term_months": {"type": "integer", "title": "Term Months"}, "vs_current_delta": {"type": "integer", "title": "Vs Current Delta"}, "headline": {"type": "string", "title": "Headline"}}, "type": "object", "required": ["type", "target_plan", "monthly_price", "term_months", "vs_current_delta", "headline"], "title": "OfferRung"}, "Page_CustomerSummary_": {"properties": {"data": {"items": {"$ref": "#/components/schemas/CustomerSummary"}, "type": "array", "title": "Data"}, "page": {"type": "integer", "title": "Page"}, "page_size": {"type": "integer", "title": "Page Size"}, "total": {"type": "integer", "title": "Total"}}, "type": "object", "required": ["data", "page", "page_size", "total"], "title": "Page[CustomerSummary]"}, "PitchGenerateRequest": {"properties": {"force": {"type": "boolean", "title": "Force", "default": false}}, "type": "object", "title": "PitchGenerateRequest"}, "PitchRead": {"properties": {"status": {"$ref": "#/components/schemas/PitchStatus"}, "text": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Text"}, "model": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Model"}, "grounding_ok": {"type": "boolean", "title": "Grounding Ok", "default": false}, "created_at": {"anyOf": [{"type": "string", "format": "date-time"}, {"type": "null"}], "title": "Created At"}}, "type": "object", "required": ["status"], "title": "PitchRead"}, "PitchStatus": {"type": "string", "enum": ["not_generated", "generating", "ready", "failed"], "title": "PitchStatus"}, "PlanRef": {"properties": {"id": {"type": "string", "title": "Id"}, "name": {"type": "string", "title": "Name"}, "speed_mbps": {"type": "integer", "title": "Speed Mbps"}, "price_myr": {"type": "integer", "title": "Price Myr"}}, "type": "object", "required": ["id", "name", "speed_mbps", "price_myr"], "title": "PlanRef"}, "TierCount": {"properties": {"plan": {"$ref": "#/components/schemas/PlanRef"}, "count": {"type": "integer", "title": "Count"}}, "type": "object", "required": ["plan", "count"], "title": "TierCount"}, "UsagePoint": {"properties": {"month": {"type": "string", "title": "Month"}, "gb": {"type": "integer", "title": "Gb"}}, "type": "object", "required": ["month", "gb"], "title": "UsagePoint"}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}, "input": {"title": "Input"}, "ctx": {"type": "object", "title": "Context"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}}}}